requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
faiss-cpu
numpy
pandas
datasets